Add new fields for mourning banner
This commit add some new fields to the mourning banner admin,
which should make the output of the banner more consistent and
controlable.

added fields:
- person_name
- person_birth_date
- person_death_date
- banner_image

also included is the skeleton for a field banner_link, which may now ben required
we should be able to use the banner_message for styled text output, once we stop
stripping any styling tags from that field.

TODO:
make the banner image set and get work as expected, there is a working
wordpress media picker for this element, but I don't think it is saving
as expected.

I am not outputting these new values to the template yet, either.